fix AppController
