+++ src/main/java/io/samlr/heiken/controller/ComputerController.java
@@ -4,9 +4,12 @@ import io.samlr.heiken.entity.Computer;
 import io.samlr.heiken.service.ComputerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @Controller
@@ -19,11 +22,11 @@ public class ComputerController {
         this.computerService = computerService;
     }
 
-    @RequestMapping(value = "/add", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
-    @ResponseBody
-    public Computer addComputer(@RequestBody Computer computer) {
-        return computerService.addComputer(computer);
-    }
+//    @RequestMapping(value = "/add", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
+//    @ResponseBody
+//    public Computer addComputer(@RequestBody Computer computer) {
+//        return computerService.addComputer(computer);
+//    }
 
     @RequestMapping(value = "/get/{id}", method = RequestMethod.GET, produces = "application/json;charset=utf-8")
     @ResponseBody
@@ -35,6 +38,46 @@ public class ComputerController {
     @ResponseBody
     public List<Computer> getAllComputers() {
         return computerService.getAllComputers();
+    }
+
+    @RequestMapping("/all_computers")
+    public String getAllComputers(Model model) {
+        model.addAttribute("computer", computerService.getAllComputers());
+
+        return "all_computers";
+    }
+
+    @PostMapping("filter")
+    public String getComputerByIp(@RequestParam String ip, ModelMap model) {
+        List<Computer> computers;
+
+        if (ip != null && !ip.isEmpty()) {
+            computers = computerService.getComputerByIp(ip);
+        } else {
+            computers = computerService.getAllComputers();
+        }
+        model.addAttribute("computers", computers);
+        return "filter";
+    }
+
+    @RequestMapping(value = "/add", method = RequestMethod.GET)
+    public String newComputer(ModelMap model){
+        Computer computer = new Computer();
+        model.addAttribute("computer", computer);
+        model.addAttribute("edit", false);
+        return "registration";
+    }
+
+    @RequestMapping(value = { "/add" }, method = RequestMethod.POST)
+    public String saveUser(@Valid Computer computer, BindingResult result,
+                           ModelMap model) {
+        if (result.hasErrors()) {
+            return "registration";
+        }
+        computerService.addComputer(computer);
+
+        model.addAttribute("success", "Computer " + computer.getArmName() + " registered successfully");
+        return "registrationSuccess";
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.PUT, produces = "application/json;charset=utf-8")
