+++ web/WEB-INF/views/registration.jsp
@@ -138,10 +138,10 @@
 			<div class="form-actions floatRight">
 				<c:choose>
 					<c:when test="${edit}">
-						<input type="submit" value="Update" class="btn btn-primary btn-sm"/> or <a href="<c:url value='/all_computers' />">Cancel</a>
+						<input type="submit" value="Update" class="btn btn-primary btn-sm"/> or <a href="<c:url value='/computer/all_computers' />">Cancel</a>
 					</c:when>
 					<c:otherwise>
-						<input type="submit" value="Register" class="btn btn-primary btn-sm"/> or <a href="<c:url value='/all_computers' />">Cancel</a>
+						<input type="submit" value="Register" class="btn btn-primary btn-sm"/> or <a href="<c:url value='/computer/all_computers' />">Cancel</a>
 					</c:otherwise>
 				</c:choose>
 			</div>
